Update Hop and use a mailbox for handling navigation changes throughout
This introduces a custom mailbox, referenced by the `linkTo` function,
which is a mailbox for string (that capture string representations of
the routes). Within our `inputs` function in `Main`, we then map over this
mailbox signal to convert the paths to Colors.App.Update.NavigateTo
actions, which get handled by the top-level reducer.